requirements.txt
Flask==3.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.8.0
requests==2.31.0
python-dotenv==1.0.0
ta==0.10.2
gunicorn==21.2.0